requirements.txt
# Telegram Bot
aiogram==3.1.1
python-telegram-bot==20.7

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# File Processing
pypdf2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
pytesseract==0.3.10
easyocr==1.7.0
openpyxl==3.1.2
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1

